requirements.txt
pytorch-lightning==1.6.0rc0
captcha
pyyaml
pillow
tqdm
wheel>=0.38.0 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools==59.5.0 # resolve https://github.com/pytorch/pytorch/issues/69894